TFKUBE-283: Move NFS module under Bitbucket module
