+++ modules/products/bitbucket/main.tf
@@ -14,7 +14,7 @@ resource "aws_route53_record" "bitbucket" {
 }
 
 module "nfs" {
-  source = "../../AWS/nfs"
+  source = "./nfs"
 
   product   = local.product_name
   namespace = var.namespace
