+++ modules/products/bitbucket/nfs/nfs-chart/templates/statefulset.yaml
